apps/api/requirements.txt
# QANTAM API Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
neo4j==5.17.0

# Redis
redis==5.0.1

# Storage
minio==7.2.3

# Search
meilisearch==0.31.0

# LLM
litellm==1.27.10
anthropic==0.18.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Document Parsing
python-docx==1.1.0
pypdf==4.0.1

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Linting
ruff==0.2.1
